--- Wordlike.java
@@ -1,10 +1,0 @@
-import java.util.Scanner;
-public class Wordlike {
-	
-	public static void main(String[] args) {
-		GUI gui = new GUI();
-		gui.runGUI();
-	}
-	
-	
-}
