+++ src/java/main/io/github/dragoncrafter10/wordlike/Wordlike.java
@@ -1,0 +1,12 @@
+package io.github.dragoncrafter10.wordlike;
+
+import java.util.Scanner;
+public class Wordlike {
+	
+	public static void main(String[] args) {
+		GUI gui = new GUI();
+		gui.runGUI();
+	}
+	
+	
+}
